Address sidebar pane review feedback
