+++ Sources/ContentView+RightSidebarCommandPalette.swift
@@ -124,23 +124,39 @@ extension ContentView {
     }
 
     static func commandPaletteRightSidebarToolPaneCommandDescriptors() -> [(mode: RightSidebarMode, commandId: String, title: String)] {
-        [
-            (
-                mode: .files,
-                commandId: "palette.openFilesPane",
-                title: String(localized: "command.openFilesPane.title", defaultValue: "Open Files as Pane")
-            ),
-            (
-                mode: .find,
-                commandId: "palette.openFindPane",
-                title: String(localized: "command.openFindPane.title", defaultValue: "Open Find as Pane")
-            ),
-            (
-                mode: .sessions,
-                commandId: "palette.openVaultPane",
-                title: String(localized: "command.openVaultPane.title", defaultValue: "Open Vault as Pane")
-            )
-        ]
+        RightSidebarMode.paneModes.compactMap { mode in
+            guard let commandId = commandPaletteRightSidebarToolPaneCommandID(mode),
+                  let title = commandPaletteRightSidebarToolPaneTitle(mode) else {
+                return nil
+            }
+            return (mode: mode, commandId: commandId, title: title)
+        }
+    }
+
+    private static func commandPaletteRightSidebarToolPaneCommandID(_ mode: RightSidebarMode) -> String? {
+        switch mode {
+        case .files:
+            return "palette.openFilesPane"
+        case .find:
+            return "palette.openFindPane"
+        case .sessions:
+            return "palette.openVaultPane"
+        case .feed, .dock:
+            return nil
+        }
+    }
+
+    private static func commandPaletteRightSidebarToolPaneTitle(_ mode: RightSidebarMode) -> String? {
+        switch mode {
+        case .files:
+            return String(localized: "command.openFilesPane.title", defaultValue: "Open Files as Pane")
+        case .find:
+            return String(localized: "command.openFindPane.title", defaultValue: "Open Find as Pane")
+        case .sessions:
+            return String(localized: "command.openVaultPane.title", defaultValue: "Open Vault as Pane")
+        case .feed, .dock:
+            return nil
+        }
     }
 
     func handleCommandPaletteRightSidebarMode(_ mode: RightSidebarMode, observedWindow: NSWindow?) {
